requirements.txt
Django==1.9.3
django-braces==1.8.1
django-dotenv==1.4.1
django-oauth-toolkit==0.10.0
djangorestframework==3.3.2
oauthlib==1.0.3
six==1.10.0
wheel==0.24.0
